+++ Backend/src/core/types/email.types.ts
@@ -1,4 +1,3 @@
-
 export interface EmailListDTO {
     _id: string;
     subject?: string;