+++ Backend/src/middlewares/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.js';
 export * from './error.handler.js';
 export * from './validator.js';
-